Correct Shell Assembler NUM pseudo-op sign bug.
Correct bug in Shell Assembler where it would not recognize a "-" sign
in the TAG field of a source card. Instruction at address 2234 in M1
was CIRA 0000, but should be CIRA 0009 to circulate the sign into the
sign digit (CIRA circulates left n+1 digits). The original sign caused
the sign digit to be lost following the shift.

The listing in the Shell Assembler manual showed CIRA 0000, but it is
possible that this was a misprint. The IBM 407 occasionally printed
a wrong character, so this is likely the cause of the error in the
transcribed program.