Preliminary run-time library for Burroughs Algebraic Compiler.
Preliminary and incomplete release of the run-time library for the
compiler, as reconstructed by Tom Sawyer. This commit is his release
dated 2022-08-18. As noted below, recovery of the library is still a
work in progress.

Tom's notes on the release from https://datatron.blogspot.com/ (slightly
edited by Paul Kimpel):
__________

The compiler is complete and authentic as distributed by Burroughs in
the early 1960s. It compiles the sample program in the manual employing
Simpson's Rule and executes with the correct answer.

  * The code for EXP and LOG are authentic originals.

  * The code for SQRT, SIN and COS are duplicates of Knuth's code found
    at Clevite Corporation.

  * The code for WRITE was painstakingly reconstructed from the
    preliminary Don Knuth code found attached to the relics utilized by
    Paul to resurrect EASY and MEASY. See
    https://archive.computerhistory.org/resources/text/Knuth_Don_X4100/
    PDF_index/k-2-pdf/k-2-c1039-ALGOL-B205.pdf

The code for READ is strictly my own work and at present will only read
integer values from cards. This is not as a great a limitation as it
might appear since the ten-digit integer 5131415927 is identical to the
result that would be loaded from 3.1415927 if it were successfully
processed by the original READ.

And here is what is still incomplete and on my current "to do" list in
priority order. I hope it won't take much more time to complete these
items, just don't look at my track record.

  * ARCTAN is not implemented.

  * READ does not handle the two types of floating point input
    xxxx.xxxx   or   xxx.xxx,nn

  * WRITE does not handle the Cardatron Punch as an output option.

  * PTREAD and PTWRITE are not implemented
__________